Create HabitatInundation code
TL/DR: 
Goal of the code is to determine the habitat types in flood plains that are inundated on various flood dates at several buffer distances from a river. I'll work to edit the code so it's a bit more universal and not just suited for my personal needs.

Overview:
The code takes the riparian/flood plain areas that were determined to be inundated on a various flood dates (0 = inundation that is connected to the river, 1 = total inundation both connected and unconnected), uses those files to clip out the habitat types in several buffer areas (habitat buffers) 
The code does the following: 
1.	Clears the data from the previous run
2.	Assigns where to output data
3.	Run data for each site (assign the site at the top of the code) 
a.	Assign location of river buffer areas (in my case there are 4)
b.	Run data for Connected data (code = 0) 
i.	Assign location of connected data to clip with the habitat buffers with inundated area.
ii.	Complete the following steps for each habitat buffer (I have 4) 
1.	Assign export path for final files
2.	Clip the first habitat inundation buffer with the  
3.	Add fields to assign site #, inundation # (1 or 2), buffer #, and the inundation date to the data you just created.
iii.	Append all the data to one common shapefile for easy exporting to Excel later. 
c.	Repeat steps in B for Connected+Unconnected data (code = 1)